+++ Front/package.json
@@ -12,7 +12,7 @@
     "axios": "^1.7.7",
     "pinia": "^2.2.4",
     "pretendard": "^1.3.9",
-    "socket.io-client": "^4.7.5",
+    "socket.io-client": "^4.8.1",
     "unplugin-auto-import": "^0.18.2",
     "unplugin-vue-components": "^0.27.4",
     "vue": "^3.4.37",
