+++ Front/src/main.js
@@ -2,7 +2,6 @@ import { createApp } from 'vue';
 import './style.css';
 import App from './App.vue';
 import router from './router/index.js';
-import pinia from './store/index.js';
 import TheHeader from './layout/TheHeader.vue';
 import BaseButton from './UI/BaseButton.vue';
 import TheFooter from './layout/TheFooter.vue';
@@ -10,16 +9,21 @@ import TheFooter from './layout/TheFooter.vue';
 import useAxios from './composables/useAxios';
 import { useSocketStore } from './store/socket';
 import { io } from 'socket.io-client';
+import { createPinia } from 'pinia';
 
 const app = createApp(App);
 
 app.use(router);
+
+const pinia = createPinia();
 app.use(pinia);
 
-// // reset axios & socket
-// const axiosInstance = useAxios();
+// reset axios & socket
+const axiosInstance = useAxios();
 
-// // local
+axiosInstance.setBaseURL('http://localhost:10111');
+
+// local
 // const socket = io('http://serverip:port'); // 이 부분만 수정 (백엔드)
 // const server = `http://${socket.io.engine.hostname}:${socket.io.engine.port}`;
 // axiosInstance.setBaseURL(server);
@@ -27,7 +31,6 @@ app.use(pinia);
 // const socketStore = useSocketStore();
 // socketStore.initSocket(socket, server);
 
-// app.vue
 app.component('the-header', TheHeader);
 app.component('the-footer', TheFooter);
 app.component('base-button', BaseButton);
