+++ Front/src/store/text.js
@@ -1,0 +1,10 @@
+import { defineStore } from 'pinia';
+
+export const useTextStore = defineStore('text', () => {
+	const text = ref('');
+	const setText = t => {
+		text.value = t;
+	};
+
+	return { text, setText };
+});
