--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "socket.io-client": "^4.8.1"
-  }
-}
